backend/requirements.txt
fastapi>=0.112
uvicorn[standard]>=0.30
SQLAlchemy>=2.0
psycopg2-binary>=2.9
pydantic>=2.6
python-dotenv>=1.0
